requirements.txt
flask-sqlalchemy
Flask
flask-migrate
flask_wtf
python-dotenv
